Refactor for smoother cmake usage
