+++ src/main_server.cpp
@@ -1,6 +1,6 @@
 #include "scheduler.hpp"
 #include "security.hpp"
-#include "server.hpp"
+#include "task_server.hpp"
 
 auto main() -> int {
     preventRootExecution();
